[IMP] OD-199, sentry: allow to distinguish between various environments in config
On Odoo.sh, newly cloned test instances are started with the unmodified
configuration file from the production environment. To allow for different
sentry settings in this case, allow to mark all sentry_ options for a
specific environment fetched from the `ODOO_STAGE` environment variable
that is present on all Odoo.sh instances.