[FIX] auditlog: adapt exclude fields test to #3137
In #3137 it is prevented to create auditlogs without log lines.

In this particular test, it is tested that no log line is created for excluded
fields. We would expect therefore that no auditlog is created after #3137, but
depending on the setup, an log line may be created for field phone_sanitized
which is not an excluded field in the test.

To fix this, the test is adapted to write another, non excluded field so that
the auditlog is created in any case.