[REF] OD-1085, auditlog: call logging with actual rule
Rather than calling logging on the rule model, then search for a rule with a
matching model, browse the rule id in the patching methods and call `create_logs`
on the rule. This makes it less awkward and more deterministic to fetch
additional info from the rule in the patching methods.